+++ csharp/src/FlatColumnsResultBuilder.cs
@@ -39,7 +39,7 @@ namespace AdbcDrivers.Databricks
             var dataTypeBuilder = new Int32Array.Builder();
             var typeNameBuilder = new StringArray.Builder();
             var columnSizeBuilder = new Int32Array.Builder();
-            var bufferLengthBuilder = new Int8Array.Builder();
+            var bufferLengthBuilder = new Int32Array.Builder();
             var decimalDigitsBuilder = new Int32Array.Builder();
             var numPrecRadixBuilder = new Int32Array.Builder();
             var nullableBuilder = new Int32Array.Builder();
@@ -72,7 +72,7 @@ namespace AdbcDrivers.Databricks
                     if (info.Precision[i].HasValue) columnSizeBuilder.Append(info.Precision[i]!.Value); else columnSizeBuilder.AppendNull();
 
                     int? bufLen = ColumnMetadataHelper.GetBufferLength(info.TypeName[i]);
-                    if (bufLen.HasValue) bufferLengthBuilder.Append((sbyte)bufLen.Value); else bufferLengthBuilder.AppendNull();
+                    if (bufLen.HasValue) bufferLengthBuilder.Append(bufLen.Value); else bufferLengthBuilder.AppendNull();
 
                     if (info.Scale[i].HasValue) decimalDigitsBuilder.Append(info.Scale[i]!.Value); else decimalDigitsBuilder.AppendNull();
